#  implement vectorized evaluation for builtinArithmeticModIntSig